+++ packages/gamut/src/index.tsx
@@ -22,7 +22,9 @@ export * from './FocusTrap';
 export * from './Form';
 export * from './GridForm';
 export * from './HiddenText';
-export * from './Typography';
+export * from './Typography/HeadingDeprecated';
+export * from './Typography/Text';
+export * from './Typography/TextDeprecated';
 export * from './HighlightedText';
 export * from './InputStepper';
 export * from './Interstitial';
