+++ packages/styleguide/stories/Typography/Text.stories.mdx
@@ -2,7 +2,7 @@ import { Column, LayoutGrid, Text } from '@codecademy/gamut/src';
 import {
   typographyElementVariants,
   typographyStyleVariants,
-} from '@codecademy/gamut/src/Typography/Text';
+} from '@codecademy/gamut/src/Typography/variants';
 import title from '@codecademy/macros/lib/title.macro';
 import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
 import { Fragment } from 'react';
